requirements-dev.txt
flake8==4.0.1
pytest-django==4.5.2
pytest-cov==3.0.0
factory-boy==3.2.1
pytest-factoryboy==2.1.0

-r requirements.txt